// src/services/types.ts
export interface RootAPIResponse<T> {
  hasError: boolean
  value: T
  error: Error
}

export interface PaginatedResult<T> {
  currentPage: number
  pageCount: number
  pageSize: number
  totalRecords: number
  result: T[]
}

export interface AdminUser {
  firstName: string
  lastName: string
  emailAddress: string
}

export interface Error {
  errorCode: string
  message: string
  isTransient: boolean
  inner: string
  fullMessage: string
}
// ...
export interface ICollection {
  amount: number
  customerId: string
  currency: string
  time: string
  customerName: string
  invoiceNumber: string
  transactionReference: string
}

export interface IOrganizationReport {
  userCount: number
  payableItemCount: number
  subOrganizationCount: number
}

export interface IFilterReport {
  currency?: string
  duration?: string
  startDate?: string
  endDate?: string
}

export interface ICollectionTrend {
  labels: string[]
  datasets: {
    label: string
    data: number[]
    maxBarThickness?: number
    backgroundColor?: string | string[]
    borderColor?: string
    tension?: number
    color?: string
  }[]
}